+++ examples/Example.js
@@ -233,7 +233,7 @@ class Example extends Component {
                 </div>
                 <div className="selection"><b>Customizing </b>
                     <UltraSelect columns={customize} rowsVisible={3} rowHeight={4} rowHeightUnit="em" onDidSelect={this.onDidSelect} titleHeight={25}
-                                 confirmButton="South Park!" getTitle={this.getTitle} getStaticText={this.getStaticText}></UltraSelect>
+                                 confirmButton="Choose" cancelButton="Give Up" getTitle={this.getTitle} getStaticText={this.getStaticText}></UltraSelect>
                 </div>
                 <div className="selection"><b>Load async data: </b>
                     <UltraSelect columns={asynColumn} onDidSelect={this.onAsync} getTitle={()=>'Language - Stargazers'}></UltraSelect></div>