+++ webpack.config.js
@@ -1,44 +1,47 @@
-var path = require("path")
-var webpack = require('webpack')
+/* global require */
+const path = require('path')
+const webpack = require('webpack')
 
-var isProd = process.env.NODE_ENV === 'production'
+const isProd = process.env.NODE_ENV === 'production'
 
-var config = {
+const config = {
     entry: {
-        "react-ultra-select": ["./src/UltraSelect.js"],
+        'react-ultra-select': ['./src/UltraSelect.js'],
     },
     externals: {
-        "react": "react",
-        "react-dom" : "react-dom",
-        "deep-equal": "deep-equal",
+        react: 'react',
+        'react-dom': 'react-dom',
+        'deep-equal': 'deep-equal',
+        iscroll: 'iscroll',
+        'iscroll-react': 'iscroll-react',
     },
     output: {
-        path: path.join(__dirname, "dist"),
-        filename: isProd ? "[name].min.js" : "[name].js",
-        library: "react-ultra-select",
-        libraryTarget: "umd"
+        path: path.join(__dirname, 'dist'),
+        filename: isProd ? '[name].min.js' : '[name].js',
+        library: 'react-ultra-select',
+        libraryTarget: 'umd',
     },
     plugins: [],
     module: {
         loaders: [
             {
                 test: /\.js$/,
-                loaders: ["babel"],
-                exclude: /node_modules/
+                loaders: ['babel'],
+                exclude: /node_modules/,
             },
             {
                 test: /\.less$/,
-                loader: "style!css!less"
-            }
-        ]
-    }
+                loader: 'style!css!less',
+            },
+        ],
+    },
 }
 
 if (isProd) {
     config.plugins.push(
         new webpack.optimize.UglifyJsPlugin({
             minimize: true,
-            comments: false
+            comments: false,
         })
     )
 }