+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Change Logs
 
+## Version 1.0.11
+
+Date: 2016-09-13
+
+- Import features from v1.1.0 to v1.0.10 (i.e. iscroll-scroll branch)
+- Add `backdropAction` prop
+
 ## Version 1.1.0
 
 Date: 2016-07-30